Move TabBarController into its own file
